+++ index.js
@@ -1,7 +1,10 @@
 module.exports = {
   "rules": {
     "at-rule-empty-line-before": [ "always", {
-      except: [ "blockless-group", "first-nested" ],
+      except: [
+        "blockless-after-same-name-blockless",
+        "first-nested",
+      ],
       ignore: ["after-comment"],
     } ],
     "at-rule-name-case": "lower",
@@ -25,8 +28,14 @@ module.exports = {
     "comment-no-empty": true,
     "comment-whitespace-inside": "always",
     "custom-property-empty-line-before": [ "always", {
-      except: [ "after-custom-property", "first-nested" ],
-      ignore: [ "after-comment", "inside-single-line-block" ],
+      except: [
+        "after-custom-property",
+        "first-nested",
+      ],
+      ignore: [
+        "after-comment",
+        "inside-single-line-block",
+      ],
     } ],
     "declaration-bang-space-after": "never",
     "declaration-bang-space-before": "always",
@@ -41,8 +50,14 @@ module.exports = {
     "declaration-colon-space-after": "always-single-line",
     "declaration-colon-space-before": "never",
     "declaration-empty-line-before": [ "always", {
-      except: [ "after-declaration", "first-nested" ],
-      ignore: [ "after-comment", "inside-single-line-block" ],
+      except: [
+        "after-declaration",
+        "first-nested",
+      ],
+      ignore: [
+        "after-comment",
+        "inside-single-line-block",
+      ],
     } ],
     "function-calc-no-unspaced-operator": true,
     "function-comma-newline-after": "always-multi-line",