+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Head
 
+- Changed: `at-rule-empty-line-before` now uses the `blockless-after-same-name-blockless` `except` option, rather than the `blockless-group` one.
 - Added: `block-closing-brace-empty-line-before` rule.
 - Added: `comment-no-empty` rule.
 - Added: `custom-property-empty-line-before` rule.